test: added TopNavigationMenu and SideNavigationMenu tests, formatted whole project code
